tests/requirements.txt
bcrypt
bs4
flask
mysql_connector
pylint
pytest
requests-html
